vonage-ac-chatbot/pyproject.toml
[project]
name = "vonage-chatbot"
version = "0.1.0"
description = "Vonage Audio Connector voice chatbot example using Pipecat + OpenAI."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pipecat-ai[openai,websocket,vonage-audio-connector,silero,runner]>=0.0.95",
  "python-dotenv",
  "loguru",
  "pydub>=0.25",
  "PyJWT>=2.8.0",
]
